backend/requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
distlib==0.3.8
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
filelock==3.13.3
Markdown==3.6
platformdirs==4.2.0
psycopg==3.2.1
psycopg-binary==3.2.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
virtualenv==20.25.1
